pyproject.toml
[project]
name = "qa-radar"
version = "0.1.0"
description = "QA/テスト自動化に関する記事を集約するMCPサーバー兼RSSアグリゲーター"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Y-Kanekoo" }]
keywords = ["qa", "testing", "rss", "mcp", "aggregator", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
]
# Phase ごとにdependenciesを段階的に追加していく:
#   Phase 1: feedparser, httpx, pyyaml ← 追加済み
#   Phase 3: feedgen
#   Phase 5: mcp, platformdirs
#   Phase 8: anthropic（optional-dependencies の `ai` extra に分離）
dependencies = [
    "feedparser>=6.0.11",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ai = []  # Phase 8 で `anthropic>=0.40` を追加

[project.scripts]
qa-radar = "qa_radar:main"

[project.urls]
Homepage = "https://github.com/Y-Kanekoo/qa-radar"
Repository = "https://github.com/Y-Kanekoo/qa-radar"
Issues = "https://github.com/Y-Kanekoo/qa-radar/issues"
Changelog = "https://github.com/Y-Kanekoo/qa-radar/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qa_radar"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "integration: integrationテスト（実フィードへの通信を伴う）",
]
filterwarnings = [
    # feedparser内部のCharacterEncodingOverride等を抑制
    "ignore::DeprecationWarning:feedparser.*",
]

[tool.coverage.run]
source = ["src/qa_radar"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
